requirements.txt
openai-whisper
torch
ffmpeg
pyaudio
ipykernel
pandas
numpy

tabula-py
JPype1
jupyter
ipywidgets

matplotlib
tqdm
langchain_community
ollama
openai
python-dotenv
pydub
AudioConverter
Flask
pandas-stubs